+++ packages/panels/lang/cs/messages.php
@@ -5,6 +5,7 @@ declare(strict_types=1);
 return [
     'create' => 'Nový/á :label',
     'edit' => 'Upravit :label',
+    'deleted' => ':label – smazáno',
     'save' => 'Uložit',
 
     /*