+++ packages/panels/resources/views/pages/edit-page.blade.php
@@ -11,4 +11,6 @@
     @foreach($relationManagers as $manager)
         @livewire($manager, ['ownerRecord' => $ownerRecord], key('rm-'.$loop->index))
     @endforeach
+
+    @include('wire-panels::pages.partials.action-modals')
 </div>